+++ apps/docs/src/content/03-components/content/color/examples/custom.tsx
@@ -1,0 +1,3 @@
+import { Color } from "@mittwald/flow-react-components";
+
+<Color color="#00c513">custom</Color>;